Reorganize files for pip distribution (untested, still need pyproject.toml)
