--- mtl/__init__.py
@@ -1,4 +1,0 @@
-from mtlLexer import mtlLexer
-from mtlParser import mtlParser
-from mtl import MTLAbstractSyntaxTreeExtractor
-from mtl import Operation, MTLFormula, to_ast
