+++ src/pytelo/mtl/psmtl2milp.py
@@ -3,30 +3,26 @@
  See license.txt file for license information.
  @author: Gustavo A. Cardona, Cristian-Ioan Vasile
 '''
+
 import gurobipy as grb
-from stl import Operation, RelOperation, STLFormula
 
-class pstl2milp(object):
-    '''Translate an STL formula to an MILP that captures partial satisfaction.'''
+from pytelo.mtl import Operation
 
-    def __init__(self, formula, ranges, vtypes=None, model=None, robust=False):
+
+class psmtl2milp(object):
+    '''
+    Translate an MTL formula to an MILP that captures partial satisfaction.
+    '''
+    def __init__(self, formula, model=None):
+        
         self.formula = formula
-        self.robust = robust
-        self.M = 1000
-        self.ranges = ranges
-        assert set(self.formula.variables()) <= set(self.ranges)
-
-        self.vtypes = vtypes
-        if vtypes is None:
-            self.vtypes = {v: grb.GRB.CONTINUOUS for v in self.ranges}
 
         self.model = model
         if model is None:
-            self.model = grb.Model('STL formula: {}'.format(formula))
+            self.model = grb.Model('MTL formula: {}'.format(formula))
 
         self.variables = dict()
         self.variables_z = dict()
-
         self.objectives = dict()
 
         self.__milp_call = {
@@ -38,13 +34,13 @@ class pstl2milp(object):
             Operation.UNTIL : self.until
         }
 
-    def translate(self, satisfaction=True): 
-        '''Translates the STL formula to MILP from time 0.'''
+    def translate(self): 
+        '''Translates the MTL formula to MILP from time 0.'''
         z = self.to_milp(self.formula)
         return z
 
     def to_milp(self, formula, t=0, depth=0, z_ancestors=None):
-        '''Generates the MILP from the STL formula.'''
+        '''Generates the MILP from the MTL formula.'''
         if depth not in self.objectives:
             self.objectives[depth] = 0
         if z_ancestors is None:
@@ -52,8 +48,8 @@ class pstl2milp(object):
         z, added, z_var = self.add_formula_variable(formula, t, depth, 
                                                     z_ancestors)
         if added:
-            self.__milp_call[formula.op](formula, z, t, depth, z_ancestors + 
-                                         [z_var])
+            self.__milp_call[formula.op](formula, z, t, depth, z_ancestors 
+                                         + [z_var])
         return z
 
     def add_formula_variable(self, formula, t, depth, z_ancestors): 
@@ -61,12 +57,15 @@ class pstl2milp(object):
         if formula not in self.variables:               
             self.variables[formula] = dict()
             self.variables_z[formula] = dict()
-        if t not in self.variables[formula]:           
+    
+        if t not in self.variables[formula]:            
             opname = Operation.getString(formula.op)
             identifier = formula.identifier()
             name = '{}_{}_{}'.format(opname, identifier, t)
+    
             if formula.op == Operation.PRED:
                 variable = self.model.addVar(vtype=grb.GRB.BINARY, name=name)
+    
             else:
                 variable = self.model.addVar(vtype=grb.GRB.CONTINUOUS,
                                              name=name, lb=0, ub=1)
@@ -75,40 +74,32 @@ class pstl2milp(object):
             self.objectives[depth] += variable_z
             self.model.update()
             self.model.addConstr(variable_z <= variable)
-
+    
             for variable_z_ancestor in z_ancestors:
                 self.model.addConstr(variable_z <= 1 - variable_z_ancestor)
+
             self.variables_z[formula][t] = variable_z
             self.variables[formula][t] = variable
             self.model.update()
+
             return variable, True, variable_z
         return self.variables[formula][t], False, self.variables_z[formula][t]
     
-    def add_state(self, state, t):
+    def add_state(self, state, t, z):
         '''Adds the `state` at time `t` as a variable.'''
         if state not in self.variables:
             self.variables[state] = dict()
+            
         if t not in self.variables[state]:
-            low, high = self.ranges[state]
-            vtype = self.vtypes[state]
-            name='{}_{}'.format(state, t)
-            v = self.model.addVar(vtype=vtype, lb=low, ub=high, name=name)
-            self.variables[state][t] = v
+            self.variables[state][t] = z
             self.model.update()
         return self.variables[state][t]
 
     def predicate(self, pred, z, t, depth, z_ancestors):
         '''Adds a predicate to the model.'''
         assert pred.op == Operation.PRED
-        v = self.add_state(pred.variable, t)
-        if pred.relation in (RelOperation.GE, RelOperation.GT):  
-            self.model.addConstr(v  - self.M * z <= pred.threshold)         
-            self.model.addConstr(v + self.M * (1 - z) >= pred.threshold)
-        elif pred.relation in (RelOperation.LE, RelOperation.LT):
-            self.model.addConstr(v + self.M * z >= pred.threshold)          
-            self.model.addConstr(v - self.M * (1 - z) <= pred.threshold)       
-        else:
-            raise NotImplementedError
+        self.add_state(pred.variable, t, z)
+        
 
     def conjunction(self, formula, z, t, depth, z_ancestors):
         '''Adds a conjunction to the model.'''
@@ -149,90 +140,14 @@ class pstl2milp(object):
         for t_ in range(a,b+1):
             z_children_left =  [self.to_milp(formula.left, t+t__, depth+1, 
                                              z_ancestors) 
-                                for t__ in range(0,t_)]
+                                             for t__ in range(0,t_)]
             z_children_right = [self.to_milp(formula.right, t+t_, depth+1, 
                                              z_ancestors)]
             z_children.append(z_children_right + sum(z_children_left))
 
         self.model.addConstr(z == grb.max_(z_children))
 
-    def pstl2lp(self, formula, t=0, optimize=True):
-        ''' It creates a linear problem from the formuale
-             that needs to be satisfied '''
-        lp = grb.Model("LP")
-        if self.robust and 'rho' not in self.ranges:
-            self.ranges['rho'] = (-grb.GRB.INFINITY, self.M - 1)
-        if self.robust:
-            rho_min, rho_max = self.ranges['rho']
-            self.rho = lp.addVar(vtype=grb.GRB.CONTINUOUS, name='rho',
-                                        lb=rho_min, ub=rho_max, obj=-1)        
-        else:
-            self.rho = 0
-        
-        formulae = self.predicate_pairs(formula, t)
-        self.lpvariable = dict()
-
-        for phi in formulae:
-            self.lpvariable[phi[0]] = dict()
-            name = '{}_{}'.format(phi[0], phi[1])
-            var = lp.addVar(vtype=grb.GRB.CONTINUOUS, lb=-10, ub=10, name=name)
-            self.lpvariable[phi[0]][phi[1]] = var
-
-            if phi[0].relation in (RelOperation.GE, RelOperation.GT):
-                lp.addConstr(var >= phi[0].threshold + self.rho)
-
-            if phi[0].relation in (RelOperation.LE, RelOperation.LT):
-                lp.addConstr(var <= phi[0].threshold - self.rho)    
-            lp.update()
-        
-        if optimize is True:
-            lp.optimize()
-
-        return lp
-        
-    def predicate_pairs(self, formula, t=0):
-        '''
-        It receives formula and time step and returns a set of the subformulae
-        that needs to be satisfied at the specific time. 
-        Note that Disjunction and eventually are special cases
-        '''
-       
-        ret = set()
-
-        if formula.op == Operation.PRED:
-            if self.variables[formula][t].x == 1:
-                ret = {(formula, t)}
-            else:
-                ret = {}
-
-        elif formula.op == Operation.AND:
-            for f in formula.children:
-                ret = ret.union(self.predicate_pairs(f, t))
-
-        elif formula.op == Operation.OR:
-            for f in formula.children:
-                if self.variables[f][t].x == 1:
-                    ret = ret.union(self.predicate_pairs(f, t))
-                    break
-
-        elif formula.op == Operation.ALWAYS:
-            f = formula.child
-            interval = range(int(formula.low), int(formula.high+1))
-            for t in interval:
-                ret = ret.union(self.predicate_pairs(f, t))
-
-        elif formula.op == Operation.EVENT:
-            f = formula.child
-            interval = range (int(formula.low), int(formula.high+1))
-            for t in interval:
-                if self.variables[f][t].x == 1:
-                    ret = ret.union(self.predicate_pairs(f, t))
-                    break
-
-        return ret
-
-
-    def hierarchical(self, model_name='model_test.lp', optimize=True):
+    def hierarchical(self, model_name='model_test.lp', optimize=True): 
         '''
         This method computes a hierarchical optimization formulation 
         (lexicografical) from root node all the way to the leaves (predicates)
@@ -251,13 +166,13 @@ class pstl2milp(object):
             self.model.setObjectiveN(-self.objectives[d], d, 
                                      priority=max_depth-d)
             self.model.update()
-        
+
         if optimize is True:
             self.model.optimize()
             self.model.write(model_name)
         return d
     
-    def ldf(self, model_name='model_test.lp', optimize=True):
+    def ldf(self, model_name='model_test.lp', optimize=True): 
         '''
         This method computes a Lowest depht first optimization formulation 
         making and increasing penalization from being far from the root node
@@ -268,11 +183,11 @@ class pstl2milp(object):
               the optimization of the problem, in case it is False it will only
               generate the objective function.
         '''
-        M2 = 30 # FIXME: computed based on formula size
+        M2 = 20 # FIXME: computed based on formula size
         reward = sum([term * M2**(-d) for d, term in self.objectives.items()])
         self.model.setObjective(reward, grb.GRB.MAXIMIZE)
         self.model.update()
-
+        
         if optimize is True:
             self.model.optimize()
             self.model.write(model_name)
@@ -291,6 +206,7 @@ class pstl2milp(object):
         '''
         self.model.setObjective(z, grb.GRB.MAXIMIZE)
         self.model.update()
+
         if optimize is True:
             self.model.optimize()
             self.model.write(model_name)
