+++ src/pytelo/stl/stl.py
@@ -11,9 +11,9 @@ import numpy as np
 from scipy.interpolate import interp1d
 from antlr4 import InputStream, CommonTokenStream
 
-from stlLexer import stlLexer
-from stlParser import stlParser
-from stlVisitor import stlVisitor
+from pytelo._internal.stlLexer import stlLexer
+from pytelo._internal.stlParser import stlParser
+from pytelo._internal.stlVisitor import stlVisitor
 
 
 class Operation(object):
