--- wmtl/__init__.py
@@ -1,4 +1,0 @@
-from wmtlLexer import wmtlLexer
-from wmtlParser import wmtlParser
-from wmtl import WMTLAbstractSyntaxTreeExtractor
-from mtl import Operation, MTLFormula, to_ast
